backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
google-generativeai==0.8.5
reportlab==4.4.0
python-dotenv==1.1.0
Pillow==11.2.1
opencv-python-headless==4.11.0.86
scipy==1.15.2
numpy==2.2.4
pandas==2.2.3
groq==0.18.0
pytesseract==0.3.13
pymongo==4.10.1
cloudinary==1.44.0
bcrypt==4.2.1
PyJWT==2.9.0
email-validator==2.2.0
